filter ChargeData and move config to new file
